Improve 'posframe-delete-frame'
We don't need to walk the frame list, because every posframe + buffer
pair is unique; there is at most one posframe for a single buffer, and
at most one buffer for a single posframe.

We also don't need to check the frame parameters, as the
posframe--frame buffer-local variable records the pertinent frame.